% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbmbtools.R
\docType{package}
\name{cbmbtools}
\alias{cbmbtools}
\alias{cbmbtools-package}
\title{cbmbtools: General purpose package containing functions and templates for microbiome analysis pipelines}
\description{
The cbmbtools package provides several categories of functions for data import, wrangling, and plotting.
}
\section{data import}{

\code{load_shared} and \code{load_tax} handle loading and basic processing of .shared and .cons.taxonomy files respectively.
}

\section{data wrangling functions}{

Some examples of data wrangling functions are \code{agg_otus}, \code{join_tax}, and \code{paste_tax}.
}

\section{plot functions}{

The plot functions currently include, \code{plot_ra}, \code{plot_spider}, and \code{plot_many_rf_varimp}.
}

