% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify.R
\name{chunkify}
\alias{chunkify}
\title{Generate R Markdown code chunks.}
\usage{
chunkify(stem, vals, chunk_opts = NULL)
}
\arguments{
\item{stem}{Sequence of unchanging characters in the code chunk label.}

\item{vals}{A vector containing the levels or values to paste together with \code{stem}.}

\item{chunk_opts}{An optional string containing any code chunk options to include in the code chunk header.}
}
\value{
No value is returned. A sequence of code chunks equal to the length of \code{vals} is output to the console. These can then be pasted into the .Rmd.
}
\description{
Generate R Markdown code chunks.
}
\examples{
chunkify(stem = "chunk_name", vals = letters[1:6], chunk_opts = "fig.height=12, fig.width=12")
}
