% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tax.R
\name{load_tax}
\alias{load_tax}
\title{Reads in a file containing taxonomy data and performs basic processing.}
\usage{
load_tax(tax_file, otu_good = NULL)
}
\arguments{
\item{tax_file}{A cons.taxonomy file resulting from processing via mothur.}

\item{otu_good}{An optional matrix containing trimmed OTU count data.}
}
\value{
Returns a dataframe containing taxonomy data.
}
\description{
Reads in a file containing taxonomy data and performs basic processing.
}
\examples{
load_tax("miseq.trim.contigs.good.unique.good.filter.unique.precluster.pick.pick.opti_mcc.shared")
load_tax("miseq.trim.contigs.good.unique.good.filter.unique.precluster.pick.pick.opti_mcc.shared", otu_good)
}
