% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hel_pca.R
\name{make_hel_pca}
\alias{make_hel_pca}
\title{Create hellinger transformed data and PCA model from \code{filt_df}. Currently assumes \code{filt_df} is subset and contains no metadata columns.}
\usage{
make_hel_pca(filt_df)
}
\arguments{
\item{filt_df}{A dataframe that has been subset such that it contains only the observations of interest and has no metadata columns.}
}
\value{
Although no value is returned, \code{filt_df} will be used to create hellinger transformed data with "_hel" appended to \code{filt_df}.
Likewise, a PCA model is created and saved to an object with "_pca" appended to \code{filt_df}.
}
\description{
Create hellinger transformed data and PCA model from \code{filt_df}. Currently assumes \code{filt_df} is subset and contains no metadata columns.
}
\examples{
otu_df_exp1 <- dplyr::filter(otu_df, Experiment == "Exp1")[,-1:-6]
make_hel_pca(otu_df_exp1)
}
