% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_rf.R
\name{many_rf}
\alias{many_rf}
\title{Generate a list of randomForest models.}
\usage{
many_rf(df, y, nforest = 100, seed = 1234, ...)
}
\arguments{
\item{df}{Dataframe that contains the variables to be included in the model.}

\item{y}{The response vector.}

\item{nforest}{Number of random forests to create, 100 by default.}

\item{seed}{Numeric value passed to \code{\link{set.seed}}, for ensuring reproducibility. Although the default is 1234, best practice would be to vary the seed, perhaps by project, or even by model.}

\item{...}{Optional parameters to pass to \code{\link[randomForest]{randomForest}}.}
}
\value{
A list object containing \code{nforest randomForest} models.
}
\description{
Generate a list of randomForest models.
}
\examples{
many_rf(data = vfd_fam.norm[, -length(vfd_fam.norm)], y = vfd_fam.norm[, "VFDandAlive"],
  importance = TRUE)
}
