% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_tax.R
\name{paste_tax}
\alias{paste_tax}
\title{Paste Genus information onto a vector of OTU numbers.}
\usage{
paste_tax(otu_vec, tax = otu.good.taxonomy)
}
\arguments{
\item{otu_vec}{Vector of OTU numbers, used to select rownames (OTUs) in taxonomy table.}

\item{tax}{Taxonomy table. Default is otu.good.taxonomy (used historically).}
}
\value{
A factor vector comprised of taxonomy label and OTU number, with levels that match the order of OTUs in \code{otu_vec}.
}
\description{
Paste Genus information onto a vector of OTU numbers.
}
\examples{
paste_tax(rownames(otu.good.taxonomy)[1:20])
otu.df$OTU <- paste_tax(otu.df$OTU)
}
