% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ra.R
\name{plot_ra}
\alias{plot_ra}
\title{Create an ordered relative abundance plot}
\usage{
plot_ra(df, title = "", error_bar = FALSE, facet_var = NULL,
  tax_fill = NULL, gtxt = FALSE)
}
\arguments{
\item{df}{Dataframe containing metadata and OTU abundance data in 'long' format, e.g. using tidyr::gather().}

\item{title}{String to use as plot title. Defaults to an empty string.}

\item{error_bar}{Boolean for whether to include an errorbar. If so, \code{df} should contain a column named SEM with this information.}

\item{facet_var}{Variable to use for faceted plots. By default, no faceting will be used.}

\item{tax_fill}{Variable to use for fill color based on taxonomy table. By default, fill color is not based on taxonomy.}

\item{gtxt}{Boolean for whether to include a geom_text() layer in the plot which will display actual percentages. Not included by default.}
}
\value{
A plot created based on \code{df} and the specified parameters.
}
\description{
Create an ordered relative abundance plot
}
\examples{
plot_ra(mock.df, title = "Cross-Comparison of Mock Control Replicates", facet_var = "Sample", tax_fill = "Phylum", txt = T)
plot_ra(samp.df, title = "Exp1 & 2 Stomachs by Treatment, Ordered by Untreated", facet_var = "Group", tax_fill = "Phylum", error_bar = T)
}
