% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spider.R
\name{plot_spider}
\alias{plot_spider}
\title{Create an ordination with spiders based on \code{spider_vec} and coloring determined by \code{col_vec}.}
\usage{
plot_spider(
  pca_rda_obj,
  spider_vec,
  col_vec,
  col_pal,
  mtitle = "",
  col_leg_pos = "bottomright",
  col_leg_title = "",
  sig_leg_pos = "topright",
  sig_leg = ""
)
}
\arguments{
\item{pca_rda_obj}{Model object resulting from a call to \code{\link[vegan]{rda}}. The proportion of variance explained by PC1 and PC2 is extracted with \code{\link{summary}}.}

\item{spider_vec}{Vector of metadata for determining spiders.}

\item{col_vec}{Vector of metadata for determining how points should be colored.}

\item{col_pal}{Palette to use for the plot's color scheme. Note that it should be at least as long as the number of unique values in \code{col_vec}.}

\item{mtitle}{The main title of the plot, defaults to an empty string.}

\item{col_leg_pos}{Position to place the color legend, defaults to the bottom right corner. This should be adjusted if it overlaps any points.}

\item{col_leg_title}{Title to be used for the color legend, defaults to an empty string, which results in no legend being shown.}

\item{sig_leg_pos}{Position to place the significance legend, defaults to the top right corner. This should be adjusted if it overlaps any points.}

\item{sig_leg}{Text describing the result of significance test, defaults to an empty string, which results in no legend being shown.}
}
\value{
A plot is displayed based on \code{pca_rda_obj} and the specified parameters.
}
\description{
Create an ordination with spiders based on \code{spider_vec} and coloring determined by \code{col_vec}.
}
\examples{
plot_spider(pca_rda_obj = otu_exp1_pca, spider_vec = otu_exp1$Treatment_group,
    col_vec = otu_exp1$Organ, col_pal = my_pal,
    mtitle = "PCA of Experiment1 Samples by Treatment Group and Organ",
    col_leg_pos = "bottomright", col_leg_title = "Organ", sig_leg_pos = "topleft",
    sig_leg = "Difference between Treatment Groups:\n permanova: p < 0.0001")
}
