% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_otus.R
\name{rank_otus}
\alias{rank_otus}
\title{Generate a ranked vector of OTUs.}
\usage{
rank_otus(df, desc = TRUE)
}
\arguments{
\item{df}{A dataframe containing OTU data to be used to determine ranking.}

\item{desc}{Boolean for determining whether the OTUs should be sorted in descending (default) or ascending order.}
}
\value{
A character vector of ranked OTUs.
}
\description{
Generate a ranked vector of OTUs.
}
\examples{
rank_otus(df = dplyr::filter(otu_df, Experiment == "Amx"))
}
