% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tax.R
\name{rank_tax}
\alias{rank_tax}
\title{Generate a ranked vector of OTUs or other taxonomic level.}
\usage{
rank_tax(df, desc = TRUE, tax_level = "OTU",
  tax_df = otu_good_taxonomy)
}
\arguments{
\item{df}{A dataframe containing OTU data to be used to determine ranking.}

\item{desc}{Boolean for determining whether the OTUs should be sorted in descending (default) or ascending order.}

\item{tax_level}{Taxonomic level in \code{tax_df} to be utilized for ranking. The values within \code{tax_level} should also be column names in \code{df}. Defaults to OTU.}

\item{tax_df}{Taxonomy table. Defaults to \code{otu_good_taxonomy}.}
}
\value{
A ranked character vector based on the mean abundance corresponding to levels in \code{tax_df[, tax_level]}.
}
\description{
Generate a ranked vector of OTUs or other taxonomic level.
}
\examples{
rank_otus(df = dplyr::filter(otu_df, Experiment == "Amx"))
rank_otus(phy_filt, tax_level = "Phylum")
}
