% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_sort_gather.R
\name{taxon_sort_gather}
\alias{taxon_sort_gather}
\title{Transform data from wide to long, join a taxonomic level, aggregate and create summary statistics, and return n OTUs.}
\usage{
taxon_sort_gather(df, n = 50, facet_var = NULL, ord_val = NULL,
  tax_df = otu_good_taxonomy, tax_level = "Phylum")
}
\arguments{
\item{df}{Dataframe containing metadata and taxon (Phylum, Otu, etc) abundance data in wide format.}

\item{n}{Number of OTUs to return, 50 by default.}

\item{facet_var}{Variable to be used for grouping and eventual faceting, possibly in \code{plot_ra}.}

\item{ord_val}{Level with the \code{facet_var} to be used for ranking OTUs.}

\item{tax_df}{Taxonomy table. Passed to \code{\link{join_tax}}. Defaults to \code{otu_good_taxonomy}.}

\item{tax_level}{Taxonomic level to join from \code{tax_df}. Defaults to Phylum.}
}
\value{
A dataframe that has been aggregated and reshaped into long form.
}
\description{
Transform data from wide to long, join a taxonomic level, aggregate and create summary statistics, and return n OTUs.
}
\examples{
filt_ord_facet_df <- dplyr::filter(otu_good_14d, Day == 0, Organ == "Feces") \%>\%
  taxon_sort_gather(facet_var = "Treatment_group", ord_val = "untreated", tax_level = "Phylum")
}
