% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_ind.R
\name{tsg_ind}
\alias{tsg_ind}
\title{Transform data from wide to long, without aggregating individual samples.}
\usage{
tsg_ind(
  df,
  n = 50,
  ar_mean = FALSE,
  geo_mean = FALSE,
  otu_ord = NULL,
  tax_df = otu_good_taxonomy,
  tax_level = NULL,
  sample_col = "Sample_name"
)
}
\arguments{
\item{df}{Dataframe containing metadata and taxon (Phylum, Otu, etc) abundance data in wide format.}

\item{n}{Number of OTUs to return, 50 by default.}

\item{ar_mean}{Boolean to indicate whether the arithmetic mean should be appended to the data.}

\item{geo_mean}{Boolean to indicate whether the geometric mean should be appended to the data.}

\item{otu_ord}{Character or factor vector of unique OTU identifiers. These are often ordered OTUs resulting from \code{\link{taxon_sort_gather}} and are used to ensure consistency in OTU ordering and factor levels across dataframes.}

\item{tax_df}{Taxonomy table. Passed to \code{\link{join_tax}}. Defaults to \code{otu_good_taxonomy}.}

\item{tax_level}{Taxonomic level(s) to join from \code{tax_df}. Defaults to NULL, in which case no join occurs.}

\item{sample_col}{Single element character vector specifying the column in \code{df} which contains Sample names. Defaults to Sample_name.}
}
\value{
An unaggregated dataframe that has been reshaped into long form.
}
\description{
Transform data from wide to long, without aggregating individual samples.
}
\examples{
mock_gg <- dplyr::filter(otu_df, Experiment == "Zymo_Mock") \%>\%
    tsg_ind(n = 20, ar_mean = T, geo_mean = T)

filt_df_ind <- dplyr::filter(otu_df, Experiment \%in\% c("A", "B"),
    Treatment_group == "TG1", Organ == "Feces", Day == 11) \%>\%
    tsg_ind(otu_ord = levels(filt_df$OTU))
}
\seealso{
\code{\link{taxon_sort_gather}}
}
