% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbmbtools.R
\docType{package}
\name{cbmbtools}
\alias{cbmbtools}
\title{cbmbtools: General purpose package containing functions and templates for microbiome analysis pipelines}
\description{
The cbmbtools package provides several categories of functions for data import, wrangling, and plotting. This workflow is primarily geared towards data resulting from processing via \href{https://www.mothur.org/}{mothur}.
}
\section{data import}{

\code{\link{load_shared}} and \code{\link{load_tax}} handle loading and basic processing of .shared and .cons.taxonomy files respectively.
}

\section{data wrangling functions}{

Some examples of data wrangling functions are \code{\link{agg_otus}}, \code{\link{join_tax}}, and \code{\link{paste_tax}}.
}

\section{plot functions}{

The plot functions currently include, \code{\link{plot_ra}}, \code{\link{plot_spider}}, and \code{\link{plot_many_rf_varimp}}.
}

