% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shared.R
\name{load_shared}
\alias{load_shared}
\title{Reads in a file containing OTU count data and keeps OTUs which meet or exceed a specified \% abundance in at least one sample.}
\usage{
load_shared(shared, otu_vec = NULL, thresh = 0.1)
}
\arguments{
\item{shared}{An opti_mcc.shared file resulting from processing via mothur.}

\item{otu_vec}{An optional vector containing OTUs to be excluded, prior to trimming. This could be obtained from extracting a subset of OTUs from the result of \code{\link{load_tax}}.}

\item{thresh}{Threshold used to determine whether any OTUs should be trimmed from the dataset. If an OTU does not reach \code{thresh} at least once in a sample, the OTU will be removed. Default threshold is 0.1.}
}
\value{
Returns a matrix containing OTU counts, (optionally) without OTUs contained in \code{otu_vec} and also trimmed to remove OTUs that did not cross the specified percentage in at least one sample.
}
\description{
Reads in a file containing OTU count data and keeps OTUs which meet or exceed a specified \% abundance in at least one sample.
}
\examples{
load_shared("miseq_mothur.shared", thresh = .05)
}
