% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tax.R
\name{load_tax}
\alias{load_tax}
\title{Reads in a file containing taxonomy data and performs basic processing.}
\usage{
load_tax(tax_file, otu_good = NULL)
}
\arguments{
\item{tax_file}{A cons.taxonomy file resulting from processing via mothur.}

\item{otu_good}{An optional matrix containing trimmed OTU count data.}
}
\value{
Returns a dataframe containing taxonomy data, optionally subset by the OTUs within `otu_good`.
}
\description{
Reads in a file containing taxonomy data and performs basic processing.
}
\examples{
load_tax("miseq_mothur.cons.taxonomy")
load_tax("miseq_mothur.cons.taxonomy", otu_good)
}
