% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_tax.R
\name{paste_tax}
\alias{paste_tax}
\title{Paste Genus information onto a vector of OTU numbers.}
\usage{
paste_tax(otu_vec, tax = otu_good_taxonomy)
}
\arguments{
\item{otu_vec}{Vector of OTU identifiers, used to select rownames (OTU IDs) in \code{tax}.}

\item{tax}{Taxonomy table, otu_good_taxonomy by default.}
}
\value{
A factor vector comprised of taxonomy label and OTU identifier, with levels that match the order of OTUs in \code{otu_vec}.
}
\description{
Paste Genus information onto a vector of OTU numbers.
}
\examples{
paste_tax(rownames(otu_good_taxonomy)[1:20])
otu_df$OTU <- paste_tax(otu_df$OTU)
}
