% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ra.R
\name{plot_ra}
\alias{plot_ra}
\title{Create an ordered relative abundance plot}
\usage{
plot_ra(df, df_obs = NULL, taxon = "OTU", yvar = "Mean_Perc",
  title = "", error_bar = FALSE, facet_var = NULL, fill = NULL,
  phy_vec = NULL, fill_pal = NULL, gtxt = FALSE, seed = 123)
}
\arguments{
\item{df}{Dataframe containing metadata and taxon (Phylum, Otu, etc) abundance data in 'long' format, e.g. using \code{\link[tidyr]{gather}}.
This dataframe is typically the result of processing via \code{\link{taxon_sort_gather}}.}

\item{df_obs}{An optional dataframe containing individual observations, when \code{df} contains summarized samples. \code{df_obs} will be used to add a points layer with \code{\link[ggplot2]{geom_jitter}}.}

\item{taxon}{Which taxonomic level within \code{df} (and optionally, \code{df_obs}) should be plotted? OTU by default.}

\item{yvar}{Variable to use on the y-axis. Defaults to \code{Mean_Perc}.}

\item{title}{String to use as plot title. Defaults to an empty string.}

\item{error_bar}{Boolean for whether to include an errorbar. If so, \code{df} should contain a column named SEM with this information.}

\item{facet_var}{Variable to use for faceted plots. By default, no faceting will be used.}

\item{fill}{Optional variable to use for fill color. Common use cases include filling by Phylum, Sample_name, or Sample_type.}

\item{phy_vec}{Optional vector containing the unique Phyla resulting from \code{\link{load_tax}}, likely following trimming. Combined with \code{fill_pal} this allows for a consistent color scheme across RA plots.}

\item{fill_pal}{An optional user-defined palette for maintaining consistent fill colors across RA plots. Best when combined with \code{phy_vec}.}

\item{gtxt}{Boolean for whether to include a \code{\link[ggplot2]{geom_text}} layer in the plot which will display actual percentages. Not included by default.}

\item{seed}{Integer to be used as a random seed to ensure reproducibility. For example, this would come into play if using \code{df_obs} which will create a \code{\link[ggplot2]{geom_jitter}} layer that adds random noise to points.}
}
\value{
A ggplot created based on \code{df} and the specified parameters. It can be further modified with additional layers.
}
\description{
Create an ordered relative abundance plot
}
\examples{
plot_ra(mock_df, title = "Cross-Comparison of Mock Control Replicates", yvar = "Percentage", facet_var = "Sample",
   fill = "Phylum", gtxt = T)

plot_ra(samp_df, title = "Exp1 & 2 Stomachs by Treatment, Ordered by Untreated",
   facet_var = "Group", fill = "Phylum", error_bar = T)

plot_ra(df = otu_agg, df_obs = otu_obs, title = "Feces Samples, Ordered by Day 0",
   facet_var = "Day", fill = "Phylum")

plot_ra(df = phy_agg, df_obs = phy_obs, taxon = "Phylum",
   title = "Feces Samples, Aggregated by Phylum, Ordered by Day 0",
   facet_var = "Day", fill = "Phylum", phy_vec = phy_rank$Phylum, fill_pal = phy_pal)
}
