% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_ind.R
\name{tsg_ind}
\alias{tsg_ind}
\title{Transform data from wide to long, without aggregating individual samples.}
\usage{
tsg_ind(
  df,
  n = 50,
  ar_mean = FALSE,
  geo_mean = FALSE,
  otu_vec = NULL,
  tax_df = otu_good_taxonomy,
  tax_level = NULL
)
}
\arguments{
\item{df}{Dataframe containing metadata and taxon (Phylum, Otu, etc) abundance data in wide format.}

\item{n}{Number of OTUs to return, 50 by default.}

\item{ar_mean}{Boolean to indicate whether the arithmetic mean should be appended to the data.}

\item{geo_mean}{Boolean to indicate whether the geometric mean should be appended to the data.}

\item{otu_vec}{Character or factor vector of OTU identifiers. These are OTUs resulting from a dataframe output by \code{\link{taxon_sort_gather}}.}

\item{tax_df}{Taxonomy table. Passed to \code{\link{join_tax}}. Defaults to \code{otu_good_taxonomy}.}

\item{tax_level}{Taxonomic level(s) to join from \code{tax_df}. Defaults to NULL, in which case no join occurs.}
}
\value{
An unaggregated dataframe that has been reshaped into long form.
}
\description{
Transform data from wide to long, without aggregating individual samples.
}
\examples{
mock_gg <- dplyr::filter(otu_df, Experiment == "Zymo_Mock") \%>\%
    tsg_ind(n = 20, ar_mean = T, geo_mean = T)

filt_df_ind <- dplyr::filter(otu_df, Experiment \%in\% c("A", "B"), Treatment_group == "TG1", Organ == "Feces", Day == 11) \%>\%
    tsg_ind()
}
\seealso{
\code{\link{taxon_sort_gather}}
}
